# docs/swagger.yaml
basePath: /api/v1
definitions:
  kicknroar_backend_internal_dto_request.CreateMatchRequest:
    description: Match creation request with venue, date, time, and player details
    properties:
      cost_per_player:
        example: 500
        minimum: 1
        type: integer
      description:
        example: Casual futsal match for intermediate players
        type: string
      duration_hours:
        example: 2
        maximum: 4
        minimum: 0.5
        type: number
      match_date:
        example: "2024-12-01T00:00:00Z"
        type: string
      match_type:
        enum:
        - casual
        - competitive
        - tournament
        example: casual
        type: string
      max_players:
        example: 10
        maximum: 22
        minimum: 2
        type: integer
      rules_notes:
        example: Bring your own water bottle
        type: string
      skill_level_required:
        enum:
        - beginner
        - intermediate
        - advanced
        - professional
        example: intermediate
        type: string
      start_time:
        example: "2024-12-01T18:00:00Z"
        type: string
      title:
        example: Weekend Futsal Match
        maxLength: 100
        minLength: 3
        type: string
      venue_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      visibility:
        enum:
        - public
        - private
        - friends_only
        example: public
        type: string
    required:
    - cost_per_player
    - duration_hours
    - match_date
    - match_type
    - max_players
    - start_time
    - title
    - venue_id
    - visibility
    type: object
...
  kicknroar_backend_internal_dto_request.UpdateMatchRequest:
    description: Match update request - all fields are optional
    properties:
      cost_per_player:
        example: 600
        minimum: 1
        type: integer
      description:
        example: Updated match description
        type: string
      duration_hours:
        example: 2.5
        maximum: 4
        minimum: 0.5
        type: number
      match_date:
        example: "2024-12-01T00:00:00Z"
        type: string
      match_type:
        enum:
        - casual
        - competitive
        - tournament
        example: competitive
        type: string
      max_players:
        example: 12
        maximum: 22
        minimum: 2
        type: integer
      rules_notes:
        example: Updated rules and notes
        type: string
      skill_level_required:
        enum:
        - beginner
        - intermediate
        - advanced
        - professional
        example: advanced
        type: string
      start_time:
        example: "2024-12-01T19:00:00Z"
        type: string
      title:
        example: Weekend Futsal Match - Updated
        maxLength: 100
        minLength: 3
        type: string
      visibility:
        enum:
        - public
        - private
        - friends_only
        example: public
        type: string
    type: object
...
  kicknroar_backend_internal_dto_response.ErrorDetail:
    description: Detailed error information
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      details:
        example:
        - email is required
        - password must be at least 8 characters
        items:
          type: string
        type: array
      message:
        example: Invalid request data
        type: string
    type: object
  kicknroar_backend_internal_dto_response.ErrorResponse:
    description: Standard error response format
    properties:
      error:
        $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorDetail'
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@kickandroar.com
    name: API Support
  description: Football match coordination platform for Dhaka, Bangladesh
  license:
    name: Proprietary
  title: Kick&Roar API
  version: "1.0"
paths:
...
  /api/v1/matches:
    get:
      consumes:
      - application/json
      description: Get a list of matches with optional filters (status, date range,
        pagination)
      parameters:
      - description: Filter by status (open, full, cancelled, completed)
        in: query
        name: status
        type: string
      - description: Filter matches from date (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: Filter matches to date (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      - description: 'Number of results per page (default: 20)'
        in: query
        name: limit
        type: integer
      - description: 'Number of results to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
      security:
      - Bearer: []
      summary: List matches
      tags:
      - matches
    post:
      consumes:
      - application/json
      description: Create a new football match with venue, date, time, and other details
      parameters:
      - description: Match creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/kicknroar_backend_internal_dto_request.CreateMatchRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new match
      tags:
      - matches
  /api/v1/matches/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a match (only creator can delete)
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete match
      tags:
      - matches
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific match
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
      security:
      - Bearer: []
      summary: Get match by ID
      tags:
      - matches
    put:
      consumes:
      - application/json
      description: Update match details (only creator can update)
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: string
      - description: Match update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/kicknroar_backend_internal_dto_request.UpdateMatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
      security:
      - Bearer: []
      summary: Update match
      tags:
      - matches
  /api/v1/matches/{id}/join:
    post:
      consumes:
      - application/json
      description: Join a match as a participant
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
      security:
      - Bearer: []
      summary: Join a match
      tags:
      - matches
  /api/v1/matches/{id}/leave:
    post:
      consumes:
      - application/json
      description: Leave a match that you previously joined
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
      security:
      - Bearer: []
      summary: Leave a match
      tags:
      - matches
  /api/v1/matches/{id}/participants:
    get:
      consumes:
      - application/json
      description: Get list of all participants for a specific match
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
      security:
      - Bearer: []
      summary: Get match participants
      tags:
      - matches
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"